Faster path descendants check
The lines in this method become a huge bottleneck once an application
reaches a size where there are many i18n locale files being loaded.
Instead of using Pathname#ascend, which iterates over all ascendants of
the given path, we can just string compare the base and other. This is
much faster and avoids scaling issues.