Redefine _read_attribute on every AR subclass
The reason we're doing this is we're copying the underlying instruction
sequences so every active record subclass gets a unique iseq and unique
inline cache for that method. We know that _read_attribute is a hot
method, so we're manually splitting the method among subclasses.

Enable this with SPECIALIZE_READER env var. It will only show
improvement with
https://github.com/ruby/ruby/compare/master...eileencodes:unbound-method-dup