Use ActiveSupport::Logger instead of BroadcastLogger
Core isn't ready for BroadcastLogger, so for now we are going to use ActiveSupport::Logger. So undo the changes in activerecord's railtie and in bootstrap.rb that utilize Broadcast Logger. And, re-introduce the #broadcast method on AS::Logger since it is needed.

These changes are just reverting changes from rails/rails@1fbd812 and rails/rails@5eaa5c0.