PoC: don't parse SQL queries when its known they are read or write queries
I really dislike `write_query?` because parsing SQL with regexp is
impossible, so it's only there to catch the most evident stuff.

We're also doing it multiple times per query.

I think we should only rely on that regexp when executing unknown
SQL, when we call `AbstractAdapter#exec_insert`, we shouldn't have to
match the SQL against a Regexp to try to guess if it's a write or not.