Maintain Rack 2 parameter parsing behaviour
When running under Rack 3, we'll proceed with the new behaviour. That
means we won't affect applications that already use Rack 3, but does
also mean we can't help an application that is simultaneously bumping
Rails and Rack together.

However, with this commit, we at least ensure we maintain compatibility
for applications that are currently still on Rack 2 -- and we can help
with the transition using deprecation warnings.

Fixes #53394