Revert "Optimize Active Job ObjectSerializer#serialize"
This reverts commit 1f8a0c06c163696618032d1001a0bdf9e2ac1ed1.

Some users are depending on the exact byte for byte stability of
the job arguments payload. The common use case seem to be job
unicity.

Changes to the format will come eventually, but first I need
to take the time to hear about use cases and find solutions
for seemless upgrades.